Build a single-stage multi-arm clinical trial for a normally distributed primary outcome

build_ss_norm() builds a single-stage multi-arm clinical trial design object assuming the primary outcome variable is normally distributed, like those returned by des_ss_norm.

build_ss_norm(
  n = rep(98, 3),
  alpha = 0.025,
  beta = 0.1,
  delta1 = 0.5,
  delta0 = 0,
  sigma = rep(1, K + 1),
  correction = "dunnett",
  power = "marginal",
  summary = F
)

Arguments

n
A numeric vector indicating the chosen value for n, the sample size in each arm. Defaults to rep(61, 3).

alpha
A numeric indicating the chosen value for α, the significance level. Defaults to 0.05.

beta
A numeric indicating the chosen value for β, used in the definition of the desired power. Defaults to 0.2.

delta1
A numeric indicating the chosen value for δ1, the 'interesting' treatment effect. Defaults to 0.5.

delta0
A numeric indicating the chosen value for δ0, the 'uninteresting' treatment effect. Defaults to 0.

sigma
A numeric vector indicating the chosen value for σ, the vector of the standard deviations of the responses in each arm. Defaults to rep(1, K + 1).

correction
A character string indicating the chosen multiple comparison correction. Can be any of "benjamini_hochberg", "benjamini_yekutieli", "bonferroni", "dunnett", "hochberg", "holm_bonferroni", "holm_sidak", "none", "sidak", and "step_down_dunnett". Defaults to "dunnett".

power
A character string indicating the chosen type of power to design the trial for. Can be one of "conjunctive", "disjunctive", and "marginal". Defaults to "marginal".

summary
A logical variable indicating whether a summary of the function's progress should be printed to the console. Defaults to F.

Value

A list of class "multiarm_des_ss_norm" containing the following elements

A tibble in the slot $opchar summarising the operating characteristics of the chosen design.

A numeric in the slot $N specifying N, the trial's total required sample size.

A numeric in the slot $gamma specifying the critical threshold for p-values, γ, below which null hypotheses would be rejected. Will be NA if correction is not a single-step testing procedure.

A numeric vector in the slot $gammaO specifying the critical thresholds for ordered p-values, γ, to use with the chosen step-wise testing procedure. Will be NA if correction is not a step-wise testing procedure.

A matrix in the slot $CovZ specifying the covariance matrix, Cov(Z), of the standardised test statistics.

A numeric vector in the slot $ratio specifying the vector of allocation ratios, r.

Each of the input variables.

See also

Examples

# The design for the default parameters
des     <- build_ss_norm()
# Modifying the number of experimental treatments and the sample size in each
# arm
des_K_n <- build_ss_norm(n = rep(100, 4))