Design a multi-stage drop-the-losers multi-arm clinical trial for a normally distributed primary outcome

des_dtl_norm() determines multi-stage drop-the-losers multi-arm clinical trial designs assuming the primary outcome variable is normally distributed. It computes required design components and returns information on key operating characteristics.

des_dtl_norm(
  Kv = c(2, 1),
  alpha = 0.025,
  beta = 0.1,
  delta1 = 0.5,
  delta0 = 0,
  sigma = 1,
  ratio = 1,
  power = "marginal",
  type = "variable",
  integer = FALSE,
  summary = FALSE
)

Arguments

Kv
A numeric vector of strictly decreasing values, indicating the chosen value for K, the number of experimental treatment arms present in each stage. Defaults to c(2, 1).

alpha
A numeric indicating the chosen value for α, the significance level (family-wise error-rate). Defaults to 0.025.

beta
A numeric indicating the chosen value for β, used in the definition of the desired power. Defaults to 0.1.

delta1
A numeric indicating the chosen value for δ1, the 'interesting' treatment effect. Defaults to 0.5.

delta0
A numeric indicating the chosen value for δ0, the 'uninteresting' treatment effect. Defaults to 0.

sigma
A numeric vector indicating the chosen values for σ0 and σ1, the standard deviations of the responses in the control and experimental arms. Must be of length 1 or 2. If of length 1, it is assumed that σ0 =σ1. Defaults to 1.

ratio
A numeric indicating the chosen value for r, the stage-wise allocation ratio to present experimental arms. Defaults to 1.

power
A character string indicating the chosen type of power to design the trial for. Can be "disjunctive" or "marginal". Defaults to "marginal".

type
A character string indicating the choice for the stage-wise sample size. Can be "variable" or "fixed". Defaults to "variable".

integer
A logical variable indicating whether the computed possible sample sizes required in each arm in each stage should be forced to be whole numbers. Defaults to FALSE. WARNING: If you set integer = TRUE and ratio != 1, obscure results can occur due to difficulties in identifying a suitable whole number sample size that meets the allocation ratio requirement.

summary
A logical variable indicating whether a summary of the function's progress should be printed to the console. Defaults to FALSE.

Value

A list, with additional class "multiarm_des_dtl_norm", containing the following elements:

A tibble in the slot $opchar summarising the operating characteristics of the identified design.

A numeric in the slot $e specifying e, the trial's critical rejection boundary for the final analysis.

A numeric in the slot $maxN specifying N, the trial's total required sample size.

A numeric in the slot $n_factor, for internal use in other functions.

A numeric in the slot $n1 specifying n1, the total sample size required in stage one of the trial.

A numeric in the slot $n10 specifying n10, the sample size required in the control arm in stage one of the trial.

Each of the input variables.

See also

Examples

# The design for the default parameters
des <- des_dtl_norm()