multiarm: Design of single- and multi-stage multi-arm clinical trials

multiarm provides functions to assist with the design of single- and multi-stage multi-arm clinical trials. In both cases the available functions allow for sample size determination, trial simulation, analytical operating characteristic calculation, and the production of several informative plots. An R Shiny graphical user interface is also provided to aid design determination. Further details on single-stage design can be found in Grayling and Wason (2020).

Getting started

You can install the latest development version of multiarm from Github with:

devtools::install_github("mjg211/multiarm")

An introductory example of how to make use of the package's core functionality can be found here. More detailed support is available in the package vignette, which can be accessed with vignette("multiarm"). For further help, please email michael.grayling@newcastle.ac.uk.

Details

In total, 31 functions are currently available. Their naming conventions are such that several character strings are joined together, separated by underscores. The first string indicates the purpose of the function (i.e., what type of calculation it performs):

build_###_###(): Build multi-arm clinical trial design objects, like those returned by the des_###_###() functions. For use when a specific design is of interest.

des_###_###(): Determine the sample size required by a particular type of multi-arm clinical trial design.

gui(): Provides a graphical user interface to design determination.

opchar_###_###(): Determine the operating characteristics (power, family-wise error-rates, etc.) of a supplied multi-arm clinical trial design, via multivariate normal integration.

plot.multiarm_des_###_###(): Produce informative plots (power, false discovery rate curves, etc.) relating to a supplied multi-arm clinical trial design.

sim_###_###(): Empirically estimate the operating characteristics (power, family-wise error-rates, etc.) of a supplied multi-arm clinical trial design, via simulation.

The second indicates the design:

###_dtl_###(): Relate to multi-stage drop-the-losers designs. See, e.g., Wason et al (2017).

###_gs_###(): Relate to group-sequential multi-arm multi-stage designs. See, e.g., Magirr et al (2012).

###_ss_###(): Relate to single-stage designs. See, e.g., Grayling and Wason (2020).

The third indicates what type of outcome the function is for:

###_###_bern(): Assume a Bernoulli distributed primary outcome.

###_###_norm(): Assume a normally distributed primary outcome.

References

Grayling MJ, Wason JMS (2020) A web application for the design of multi-arm clinical trials. BMC Cancer 20:80. DOI: 10.1186/s12885-020-6525-0. PMID: 32005187.

Magirr D, Jaki T, Whitehead J (2012) A generalized Dunnett test for multi-arm multi-stage clinical studies with treatment selection. Biometrika 99(2):494--501. DOI: 10.1093/biomet/ass002.

Wason J, Stallard N, Bowden J, Jennison C (2017) A multi-stage drop-the-losers design for multi-arm clinical trials. Stat Meth Med Res 26(1):508--524. DOI: 10.1177/0962280214550759. PMID: 25228636.