Analytically determine the operating characteristics of a multi-stage group-sequential multi-arm clinical trial for a normally distributed primary outcome

opchar_gs_norm() determines the operating characteristics of a specified multi-stage group-sequential multi-arm clinical trial design assuming the primary outcome variable is normally distributed, for given values of the true treatment effects, using multivariate normal integration.

opchar_gs_norm(des = des_gs_norm(), tau, summary = FALSE)

Arguments

des
A list of class "multiarm_des_gs_norm", as returned by build_gs_norm or des_gs_norm (i.e., a multi-stage group-sequential multi-arm clinical trial design for a normally distributed outcome). Defaults to des_gs_norm().

tau
A matrix whose rows indicate values of τ at which to evaluate the operating characteristics. Defaults internally to the global null, global alternative, and each of the least favourable configurations, for the specified design des.

summary
A logical variable indicating whether a summary of the function's progress should be printed to the console. Defaults to FALSE.

Value

A list containing the following elements

A tibble in the slot $opchar giving the determined operating characteristics.

A tibble in the slot $pmf_N giving the determined probability mass functions of the random required sample size.

Each of the input variables.

See also

Examples

# The operating characteristics for the default parameters
opchar        <- opchar_gs_norm()