Plot operating characteristics of a multi-stage group-sequential multi-arm clinical trial for a normally distributed primary outcome

plot.multiarm_des_gs_norm() produces power curve plots for a specified multi-stage group-sequential multi-arm clinical trial design assuming the primary outcome is normally distributed.

# S3 method for multiarm_des_gs_norm
plot(
  x = des_gs_norm(),
  delta_min = -x$delta1,
  delta_max = 2 * x$delta1,
  delta = x$delta1 - x$delta0,
  density = 100,
  output = FALSE,
  print_plots = TRUE,
  summary = FALSE,
  ...
)

Arguments

x
A list of class "multiarm_des_gs_norm", as returned by build_gs_norm or des_gs_norm (i.e., a multi-stage group-sequential multi-arm clinical trial design for a normally distributed outcome). Defaults to des_gs_norm().

delta_min
A numeric specifying the chosen minimum value for the treatment effects to include on the produced plots. Defaults to -x$delta1.

delta_max
A numeric specifying the chosen maximum value for the treatment effects to include on the produced plots. Defaults to 2*x$delta1.

delta
A numeric specifying the chosen treatment effect shift to use in the 'shifted treatment effects plot'. Defaults to x$delta1 - x$delta0.

density
A numeric variable indicating the number of treatment effect scenarios to consider for each power curve. Increasing density increases the smoothness of the produced plots, at the cost of increased run time. Defaults to 100.

output
A logical variable indicating whether the available outputs from the function (see below) should be returned. Defaults to FALSE.

print_plots
A logical variable indicating whether to print produced plots. Defaults to TRUE.

summary
A logical variable indicating whether a summary of the function's progress should be printed to the console. Defaults to FALSE.

...
Not currently used.

Value

If output = T, a list containing the following elements

A list in the slot $plots containing the produced plots.

Each of the input variables.

See also

Examples

if (FALSE) {
# The design for the default parameters
des        <- des_gs_norm()
plot(des)
}